Edit UI and Add Switch Account Logic

+++ lib/features/profile/delete_account_and_logout/delete_account/ui/delete_account_row.dart
@@ -4,6 +4,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:souq/core/helpers/extensions.dart';
 import 'package:souq/core/routing/app_routes.dart';
 import 'package:souq/core/styles/app_text_styles.dart';
+import 'package:souq/core/widgets/app_dialog.dart';
 import 'package:souq/features/profile/delete_account_and_logout/delete_account/logic/cubit/delete_account_cubit.dart';
 import 'package:souq/features/profile/delete_account_and_logout/delete_account/logic/cubit/delete_account_state.dart';
 
@@ -12,38 +13,44 @@ class DeleteAccountRow extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocListener<DeleteAccountCubit, DeleteAccountState>(
+    return BlocConsumer<DeleteAccountCubit, DeleteAccountState>(
       listener: (context, state) {
         if (state is DeleteAccountError) {
-          print('Error: ${state.message}');
-          // show snackbar or dialog
+          ScaffoldMessenger.of(
+            context,
+          ).showSnackBar(SnackBar(content: Text(state.message)));
         }
         if (state is DeleteAccountSuccess) {
           context.pushReplacementNamed(AppRoutes.signupScreen);
         }
       },
-      child: BlocBuilder<DeleteAccountCubit, DeleteAccountState>(
-        builder: (context, state) {
-          return InkWell(
-            onTap: () async {
+      builder: (context, state) {
+        return InkWell(
+          onTap: () async {
+            final shouldDelete = await showDialog<bool>(
+              context: context,
+              builder: (context) {
+                return AppDialog(
+                  desc: 'Delete',
+                  title: "Are you sure you want to\nDelete this account?",
+                );
+              },
+            );
+            if (shouldDelete == true && context.mounted) {
               await context.read<DeleteAccountCubit>().deleteAccount();
-            },
-            child: Row(
-              children: [
-                Text('Delete Account', style: AppTextStyles.redColor20FontText),
-                Spacer(),
-                state is DeleteAccountLoading
-                    ? CircularProgressIndicator()
-                    : Icon(
-                        Icons.delete_forever,
-                        size: 32.sp,
-                        color: Colors.grey,
-                      ),
-              ],
-            ),
-          );
-        },
-      ),
+            }
+          },
+          child: Row(
+            children: [
+              Text('Delete Account', style: AppTextStyles.redColor20FontText),
+              Spacer(),
+              state is DeleteAccountLoading
+                  ? CircularProgressIndicator()
+                  : Icon(Icons.delete_forever, size: 32.sp, color: Colors.grey),
+            ],
+          ),
+        );
+      },
     );
   }
 }
